Desarrollo director

+++ src/Redux/actions/pfActions.js
@@ -1,5 +1,8 @@
 import axios from 'axios';
 import { URI } from '../config';
+import { useNavigate } from 'react-router-dom';
+import { handleError, isTokenExpired } from './errorActions';
+import { logout } from './authActions';
 
 export const PREGUNTAS_FRECUENTES_REQUEST = 'PREGUNTAS_FRECUENTES_REQUEST';
 export const PREGUNTAS_FRECUENTES_SUCCESS = 'PREGUNTAS_FRECUENTES_SUCCESS';
@@ -8,11 +11,16 @@ export const CREATE_PREGUNTA_FRECUENTE_SUCCESS = 'CREATE_PREGUNTA_FRECUENTE_SUCC
 export const UPDATE_PREGUNTA_FRECUENTE_SUCCESS = 'UPDATE_PREGUNTA_FRECUENTE_SUCCESS';
 export const DELETE_PREGUNTA_FRECUENTE_SUCCESS = 'DELETE_PREGUNTA_FRECUENTE_SUCCESS';
 
-export const fetchPreguntasFrecuentes = () => async (dispatch, getState) => {
+export const fetchPreguntasFrecuentes = () => async (dispatch) => {
+  if (isTokenExpired()) {
+    dispatch(logout());
+    return;
+  }
+  
   dispatch({ type: PREGUNTAS_FRECUENTES_REQUEST });
 
   try {
-    const token = getState().auth.token;
+    const token = localStorage.getItem('token');
 
     const response = await axios.get(`${URI}/api/faq`, {
       headers: {
@@ -26,19 +34,21 @@ export const fetchPreguntasFrecuentes = () => async (dispatch, getState) => {
     });
   } catch (error) {
     console.error('Error fetching preguntas frecuentes:', error);
-    dispatch({
-      type: PREGUNTAS_FRECUENTES_FAILURE,
-      error: error.response ? error.response.data : { message: error.message }
-    });
+    dispatch(handleError(PREGUNTAS_FRECUENTES_FAILURE, error, useNavigate()));
   }
 };
 
 // Acción para crear una nueva pregunta frecuente
-export const createPreguntaFrecuente = (nuevaPregunta) => async (dispatch, getState) => {
+export const createPreguntaFrecuente = (nuevaPregunta) => async (dispatch) => {
+  if (isTokenExpired()) {
+    dispatch(logout());
+    return;
+  }
+
   dispatch({ type: PREGUNTAS_FRECUENTES_REQUEST });
 
   try {
-    const token = getState().auth.token;
+    const token = localStorage.getItem('token');;
 
     const response = await axios.post(`${URI}/api/faq`, nuevaPregunta, {
       headers: {
@@ -54,19 +64,21 @@ export const createPreguntaFrecuente = (nuevaPregunta) => async (dispatch, getSt
     dispatch(fetchPreguntasFrecuentes());
   } catch (error) {
     console.error('Error creating pregunta frecuente:', error);
-    dispatch({
-      type: PREGUNTAS_FRECUENTES_FAILURE,
-      error: error.response ? error.response.data : { message: error.message }
-    });
+    dispatch(handleError(PREGUNTAS_FRECUENTES_FAILURE, error, useNavigate()));
   }
 };
 
 // Acción para actualizar una pregunta frecuente
-export const updatePreguntaFrecuente = (id, preguntaActualizada) => async (dispatch, getState) => {
+export const updatePreguntaFrecuente = (id, preguntaActualizada) => async (dispatch) => {
+  if (isTokenExpired()) {
+    dispatch(logout());
+    return;
+  }
+
   dispatch({ type: PREGUNTAS_FRECUENTES_REQUEST });
 
   try {
-    const token = getState().auth.token;
+    const token = localStorage.getItem('token');
 
     const response = await axios.put(`${URI}/api/faq/${id}`, preguntaActualizada, {
       headers: {
@@ -82,19 +94,21 @@ export const updatePreguntaFrecuente = (id, preguntaActualizada) => async (dispa
     dispatch(fetchPreguntasFrecuentes());
   } catch (error) {
     console.error('Error updating pregunta frecuente:', error);
-    dispatch({
-      type: PREGUNTAS_FRECUENTES_FAILURE,
-      error: error.response ? error.response.data : { message: error.message }
-    });
+    dispatch(handleError(PREGUNTAS_FRECUENTES_FAILURE, error, useNavigate()));
   }
 };
 
 // Acción para eliminar una pregunta frecuente
-export const deletePreguntaFrecuente = (id) => async (dispatch, getState) => {
+export const deletePreguntaFrecuente = (id) => async (dispatch) => {
+  if (isTokenExpired()) {
+    dispatch(logout());
+    return;
+  }
+
   dispatch({ type: PREGUNTAS_FRECUENTES_REQUEST });
 
   try {
-    const token = getState().auth.token;
+    const token = localStorage.getItem('token');
 
     await axios.delete(`${URI}/api/faq/${id}`, {
       headers: {
@@ -108,19 +122,21 @@ export const deletePreguntaFrecuente = (id) => async (dispatch, getState) => {
     dispatch(fetchPreguntasFrecuentes());
   } catch (error) {
     console.error('Error deleting pregunta frecuente:', error);
-    dispatch({
-      type: PREGUNTAS_FRECUENTES_FAILURE,
-      error: error.response ? error.response.data : { message: error.message }
-    });
+    dispatch(handleError(PREGUNTAS_FRECUENTES_FAILURE, error, useNavigate()));
   }
 };
 
 // Acción para obtener preguntas frecuentes por categoría
-export const fetchPreguntasFrecuentesByCategoria = (categoriaId) => async (dispatch, getState) => {
+export const fetchPreguntasFrecuentesByCategoria = (categoriaId) => async (dispatch) => {
+  if (isTokenExpired()) {
+    dispatch(logout());
+    return;
+  }
+  
   dispatch({ type: PREGUNTAS_FRECUENTES_REQUEST });
 
   try {
-    const token = getState().auth.token;
+    const token = localStorage.getItem('token');
 
     const response = await axios.get(`${URI}/api/faq/categoria/${categoriaId}`, {
       headers: {
@@ -133,9 +149,6 @@ export const fetchPreguntasFrecuentesByCategoria = (categoriaId) => async (dispa
     });
   } catch (error) {
     console.error('Error fetching preguntas frecuentes by categoria:', error);
-    dispatch({
-      type: PREGUNTAS_FRECUENTES_FAILURE,
-      error: error.response ? error.response.data : { message: error.message }
-    });
+    dispatch(handleError(PREGUNTAS_FRECUENTES_FAILURE, error, useNavigate()));
   }
 };
